refactor: 데이터베이스 수정
