+++ core/src/test/java/kitchenpos/ApplicationTest.java
@@ -1,9 +1,10 @@
 package kitchenpos;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootApplication
 class ApplicationTest {
     @Test
     void contextLoads() {
