+++ core/src/test/java/kitchenpos/core/application/OrderServiceTest.java
@@ -1,17 +1,17 @@
-package kitchenpos.application;
+package kitchenpos.core.application;
 
-import kitchenpos.Product.repository.ProductRepository;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.repository.MenuProductRepository;
-import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.presentation.dto.OrderCreateRequest;
-import kitchenpos.order.presentation.dto.OrderLineItemCreateRequest;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.core.Product.repository.ProductRepository;
+import kitchenpos.core.menu.domain.Menu;
+import kitchenpos.core.menu.repository.MenuProductRepository;
+import kitchenpos.core.menu.repository.MenuRepository;
+import kitchenpos.core.menugroup.domain.MenuGroup;
+import kitchenpos.core.menugroup.repository.MenuGroupRepository;
+import kitchenpos.core.order.domain.Order;
+import kitchenpos.core.order.presentation.dto.OrderCreateRequest;
+import kitchenpos.core.order.presentation.dto.OrderLineItemCreateRequest;
+import kitchenpos.core.order.repository.OrderRepository;
+import kitchenpos.core.table.domain.OrderTable;
+import kitchenpos.core.table.repository.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
